# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UniVerse — Smart university platform for the Faculty of Agriculture
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List teaching staff
      parameters:
        - in: query
          name: department
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    get:
      operationId: getStaffMember
      tags: [staff]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Staff detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /quizzes/{id}/submit:
    post:
      operationId: submitQuiz
      tags: [quizzes]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitQuizBody"
      responses:
        "200":
          description: Quiz attempt result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
...
  /missions/{id}/complete:
    post:
      operationId: completeMission
      tags: [missions]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated mission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    StaffMember:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        title: { type: string }
        role: { type: string, enum: [dean, doctor, ta] }
        department: { type: string }
        bio: { type: string }
        avatarUrl: { type: ["string", "null"] }
        email: { type: string }
        officeHours: { type: ["string", "null"] }
        researchInterests:
          type: array
          items: { type: string }
      required: [id, name, title, role, department, bio, email, researchInterests]
...
    QuizSummary:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        courseId: { type: integer }
        courseTitle: { type: string }
        durationMinutes: { type: integer }
        questionCount: { type: integer }
        totalPoints: { type: integer }
        difficulty: { type: string, enum: [easy, medium, hard] }
      required: [id, title, description, courseId, courseTitle, durationMinutes, questionCount, totalPoints, difficulty]
    QuizQuestion:
      type: object
      properties:
        id: { type: integer }
        text: { type: string }
        options:
          type: array
          items: { type: string }
        points: { type: integer }
      required: [id, text, options, points]
    Quiz:
      allOf:
        - $ref: "#/components/schemas/QuizSummary"
        - type: object
          properties:
            questions:
              type: array
              items: { $ref: "#/components/schemas/QuizQuestion" }
          required: [questions]
    QuizAnswer:
      type: object
      properties:
        questionId: { type: integer }
        selectedIndex: { type: integer }
      required: [questionId, selectedIndex]
    SubmitQuizBody:
      type: object
      properties:
        answers:
          type: array
          items: { $ref: "#/components/schemas/QuizAnswer" }
      required: [answers]
    QuizQuestionResult:
      type: object
      properties:
        questionId: { type: integer }
        correct: { type: boolean }
        correctIndex: { type: integer }
        explanation: { type: string }
      required: [questionId, correct, correctIndex, explanation]
    QuizResult:
      type: object
      properties:
        quizId: { type: integer }
        score: { type: integer }
        totalPoints: { type: integer }
        correctCount: { type: integer }
        questionCount: { type: integer }
        pointsEarned: { type: integer }
        breakdown:
          type: array
          items: { $ref: "#/components/schemas/QuizQuestionResult" }
      required: [quizId, score, totalPoints, correctCount, questionCount, pointsEarned, breakdown]
...
    Mission:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        points: { type: integer }
        kind: { type: string, enum: [study, quiz, forum, attendance, focus] }
        completed: { type: boolean }
      required: [id, title, description, points, kind, completed]
...